requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.4
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
